Feat(professor): Update with definitive role assignment #15

+++ coup-de-maths/src/main/java/com/simplon/coupdemaths/repositories/securiry/OwnerRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 
-public interface OwnerRepository extends CrudRepository<OwnerRepositoryModel, Integer> {
+public interface OwnerRepository extends CrudRepository<OwnerRepositoryModel, Long> {
 //    OwnerRepositoryModel findByLogin(String login);
     OwnerRepositoryModel findByEmail(String email);
 }
